+++ Example/iamport-ios/View/PaymentMobileWebMode.swift
@@ -3,6 +3,8 @@ import SwiftUI
 import UIKit
 import WebKit
 import iamport_ios
+import RxSwift
+import RxCocoa
 
 
 struct PaymentMobileViewModeView: UIViewControllerRepresentable {
@@ -18,12 +20,13 @@ struct PaymentMobileViewModeView: UIViewControllerRepresentable {
 
 }
 
-class PaymentMobileViewModeViewController: UIViewController, WKNavigationDelegate {
+class PaymentMobileViewModeViewController: UIViewController, WKUIDelegate, WKNavigationDelegate {
+    var disposeBag = DisposeBag()
+    let webViewDelegate = MyWKWebViewDelegate()
 
     private lazy var wkWebView: WKWebView = {
         var view = WKWebView()
         view.backgroundColor = UIColor.clear
-        view.navigationDelegate = self
         return view
     }()
 
@@ -70,14 +73,21 @@ class PaymentMobileViewModeViewController: UIViewController, WKNavigationDelegat
     override func viewDidDisappear(_ animated: Bool) {
         super.viewDidDisappear(animated)
         print("PaymentMobileViewModeView viewDidDisappear")
+        disposeBag = DisposeBag()
         Iamport.shared.close()
     }
 
     private func openWebView() {
         print("오픈! 샘플 웹뷰")
 
-        let bundle = Bundle(for: type(of: self))
+        /*
+          setup I'mport WKUIDelegate & WKNavigationDelegate
+          url 을 통해 업데이트 하는 로직이 있을 경우에 [IamPortWKWebViewDelegate] 사용
+         */
+        wkWebView.uiDelegate = webViewDelegate as? WKUIDelegate
+        wkWebView.navigationDelegate = webViewDelegate as? WKNavigationDelegate
 
+        let bundle = Bundle(for: type(of: self))
         guard let url = bundle.url(forResource: "mobileweb", withExtension: "html") else {
             print("html file url 비정상")
             return
@@ -89,12 +99,35 @@ class PaymentMobileViewModeViewController: UIViewController, WKNavigationDelegat
         }
     }
 
+
     // 아임포트 SDK 결제 요청
     func requestPayment() {
         openWebView()
+
+        // IamPortWKWebViewDelegate 또는 아래의 Iamport.shared.updateWebViewUrl 을 통해 변경되는 url 을 체크 가능합니다.
+        Iamport.shared.updateWebViewUrl.subscribe { [weak self] url in
+            print("updateWebViewUrl received url : \(url.element)")
+        }.disposed(by: disposeBag)
+
         Iamport.shared.pluginMobileWebSupporter(mobileWebMode: wkWebView)
     }
 
+}
+
+
+/**
+ url 을 통해 업데이트 하는 로직이 있을 경우에 [IamPortWKWebViewDelegate] 사용하시거나,
+ 또는 Iamport.shared.updateWebViewUrl 의 subscribe 을 통해 변경되는 url 을 체크 가능합니다.
+ */
+class MyWKWebViewDelegate: IamPortWKWebViewDelegate {
+    override func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        if let url = navigationAction.request.url {
+            // TODO : write your logic
+            print("MyWKNavigationDelegate received url : \(url)")
+        }
+
+        super.webView(webView, decidePolicyFor: navigationAction, decisionHandler: decisionHandler)
+    }
 }
 
 struct PaymentMobileViewModeView_Previews: PreviewProvider {
